tqftpserv: Compilation warnings break strict builds, fix -Wall -Werror issues
Resolve multiple compilation issues triggered by -Wall and -Werror flags:

- Multiple warnings in in zstd_decompress_file function in zstd-decompress.h.
- Eliminate redefinition of zstd_decompress_file between zstd-decompress.h
  and zstd-decompress.c.
- Address pointer type mismatch in tqftpserv.c by correcting char * vs uint8_t *
  assignment.
- Resolve signedness comparison issues in tqftpserv.c for send_len and tsize
  checks.
- Remove unused parameters argc and argv in main() function of tqftpserv.c.

These changes ensure clean compilation and adherence to strict warning policies.

+++ zstd-decompress.c
@@ -85,7 +85,7 @@ int zstd_decompress_file(const char *filename)
 		return -1;
 	}
 
-	if (write(output_file_fd, decompressed_buffer, decompressed_size) != decompressed_size) {
+	if (write(output_file_fd, decompressed_buffer, decompressed_size) != (ssize_t)decompressed_size) {
 		perror("write failed");
 		close(output_file_fd);
 		free(decompressed_buffer);
